Adds and refactors tests
